# dto/dto.go
package dto

type Response struct {
	Status  int    `json:"status"`
	Message string `json:"message"`
}

type AccountResponse struct {
	Token string `json:"token"`
}

type ErrorResponse struct {
	Error string `json:"error"`
}

type RegisterPayload struct {
	Name     string `json:"name" validate:"required"`
	Email    string `json:"email" validate:"required,email"`
	Password string `json:"password" validate:"required"`
}